api: add sam3_set_image_file() and fix review issues
Add public convenience API sam3_set_image_file() that loads an image
from disk, letterboxes to model input size, and calls sam3_set_image().
Fix integer overflow in letterbox aspect ratio calculation and add
fopen error checking in test BMP helper.

+++ src/util/image.c
@@ -107,10 +107,10 @@ enum sam3_error sam3_image_letterbox(const struct sam3_image *src,
 	int sw, sh;
 	if (src->width >= src->height) {
 		sw = target_size;
-		sh = (src->height * target_size + src->width / 2) / src->width;
+		sh = (int)(((size_t)src->height * target_size + src->width / 2) / src->width);
 	} else {
 		sh = target_size;
-		sw = (src->width * target_size + src->height / 2) / src->height;
+		sw = (int)(((size_t)src->width * target_size + src->height / 2) / src->height);
 	}
 
 	/* Resize to scaled dimensions */
